appengine/requirements.txt
bottle==0.12.16
PyJWT==1.6.1
google-cloud-datastore==1.7.3
google-api-python-client==1.7.7
numpy==1.16.0